requirements.txt
git+https://github.com/apache/airflow.git@v2-8-test
apache-airflow-providers-amazon[s3fs]
apache-airflow-providers-common-io
pandas
praw
